requirements.txt
streamlit>=1.28.0
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
transformers>=4.30.0,<4.45.0
Pillow>=9.5.0,<11.0.0
numpy>=1.24.0,<2.0.0
sqlite-vec>=0.1.1
tqdm>=4.65.0
sentencepiece>=0.1.98 